Moves GenerateOrderInvoice under Order/Invoice namespace
